requirements.txt
pandas
scikit-learn
matplotlib
seaborn
kaggle
python-dotenv